Avoid duplicating crate download code
